requirements.txt
chromadb==0.4.15
sentence-transformers==2.2.2
langchain==0.0.325
langchain-community==0.0.6
openai==1.3.7
python-dotenv==1.0.0
tiktoken==0.5.1
streamlit==1.28.1
